Improve Grammars
Lexical Grammars have been updated a bit to be a bit more consistent and to better match with the actual Syntatical Grammar, and also to be more consistent with implementations.

Syntatical Grammar has been updated mostly in formatting and organization, although a few rules have been changed to better suit the implementation at this time.