Finish Parser.
Parser now fully works, and is not whitespace dependent. This gives us a bit more flexibility for changes in the language in the future. Whitespace emission is enabled by default, as it's more accurate as some 'guesses' have to be taken in non-whitespace mode.

SVSM can successfully, and properly, parse a sample of the example in SVSM through a provided integration test.