Map is now a map, add OrderedFloat crate for maps.
Map is now an actual map type in the underlying code. This should allow us to be faster, and it cleans up the code a bit and allows us to remove some custom types we had.

OrderedFloat has now been added, this allows us to actually use Floats for BTreeMaps and the like.