Clean up code to remove warnings, rename functions, and adjust main tests.
Code has now been cleaned to remove warnings that we could avoid.

Tests in lib.rs have been adjusted to not be nested, so integration_tests are a module in the lib.rs file, instead of being in the tests module.

new_vexprs() has been renamed to new_vector_ast() as the old name wasn't great and could be misread.