Update parser.rs to make sure lists and maps no longer need trailing delimiters.
For a while lists and maps didn't technically need trailing delimiters (',' for list, and ';' for maps). While the syntax still technically requires it, the parser does not actually enforce it. However, the parser would not properly parse symbols without the trailing delimiters in some circumstances (it would instead parse as a function call. Which is incorrect). As such Parser should now properly parse in those circumstances.

I do intend on updating the syntax to denote this relaxation, but for now it's just unofficial.